# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Family Wallet API",
    "description": "API для управління спільним сімейним бюджетом, відстеження транзакцій та організації таємних зборів (Secret Gift Mode).",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/groups": {
      "get": {
        "tags": [
          "Groups"
        ],
        "summary": "Мої групи",
        "description": "Повертає список груп, у яких користувач є учасником.",
        "operationId": "get_my_groups_api_v1_groups_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "401": {
            "description": "Помилка (Validation, Access, Not Found or Conflict)",
            "content": {
              "application/json": {
                "example": {
                  "timestamp": "2026-04-09T10:00:00Z",
                  "errorCode": "SPECIFIC_ERROR_CODE",
                  "message": "Опис проблеми зрозумілою мовою",
                  "path": "/api/v1/...",
                  "details": []
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Groups"
        ],
        "summary": "Створити групу",
        "description": "Створює нову сімейну групу та генерує інвайт-лінк.",
        "operationId": "create_group_api_v1_groups_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GroupCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Родина Петренків"
            ]
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "GroupCreate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}